Refactored entity, entity instance and form
Fields are now responsible for mutating attributes before they are set on model,
also they are responsible for copying and checking whether the attribute is changed.

It's now possible to provide more presentational links (like `Printed version`) in the
schema (`exportPresentationActions`); those links're moved to the top menu of the form.

Refactored inline forms to make them less memory leaky.